requirements.txt
kokoro
soundfile
langchain-community
langchain-groq
python-dotenv